Move metrics tracker from plugin reporting to shared metrics
Relocate createMetricsTracker from src/plugin/src/reporting/metrics-tracker.js to src/shared/metrics/metrics-tracker.js so the generic instrumentation lives with other shared utilities and can be reused without crossing plugin boundaries. Update plugin imports and move the unit test under src/shared/tests to follow the new module location.
