+++ src/plugin/src/project-index/metrics.js
@@ -1,5 +1,5 @@
 import { isObjectLike } from "../../../shared/object-utils.js";
-import { createMetricsTracker } from "../reporting/metrics-tracker.js";
+import { createMetricsTracker } from "../../../shared/metrics/metrics-tracker.js";
 
 const REQUIRED_METRIC_METHODS = [
     "startTimer",
